+++ packages/marathon/extra/marathon.sh
@@ -11,6 +11,8 @@ export LIBPROCESS_IP=$($MESOS_IP_DISCOVERY_COMMAND)
 : ${MARATHON_TASK_LAUNCH_TIMEOUT=86400000}
 : ${MARATHON_TASK_RESERVATION_TIMEOUT=1800000}
 : ${MARATHON_DECLINE_OFFER_DURATION=300000}
+: ${STATSD_UDP_HOST="localhost"}
+: ${STATSD_UDP_PORT="61825"}
 : ${MARATHON_ENABLE_FEATURES="vips,task_killing,external_volumes,gpu_resources"}
 : ${MARATHON_MESOS_AUTHENTICATION_PRINCIPAL="dcos_marathon"}
 : ${MARATHON_MESOS_USER="root"}
@@ -38,6 +40,8 @@ export \
   MARATHON_TASK_LAUNCH_CONFIRM_TIMEOUT \
   MARATHON_TASK_LAUNCH_TIMEOUT \
   MARATHON_TASK_RESERVATION_TIMEOUT \
+  STATSD_UDP_HOST \
+  STATSD_UDP_PORT \
   MARATHON_ZK_COMPRESSION
 
 # TODO (DCOS_OSS-3592) - move this variable to the exported list after MARATHON-8254 is fixed, and remove --default_accepted_resources_roles below
@@ -51,4 +55,5 @@ exec $PKG_PATH/marathon/bin/marathon \
     -J-XX:+PrintGCTimeStamps \
     --master zk://zk-1.zk:2181,zk-2.zk:2181,zk-3.zk:2181,zk-4.zk:2181,zk-5.zk:2181/mesos \
     --mesos_leader_ui_url "/mesos" \
+    --metrics_statsd --metrics_statsd_host "$STATSD_UDP_HOST" --metrics_statsd_port "$STATSD_UDP_PORT" \
     --default_accepted_resource_roles "${MARATHON_DEFAULT_ACCEPTED_RESOURCE_ROLES}"