+++ CHANGES.md
@@ -41,6 +41,8 @@ This change also aligned the authentication architectures between DC/OS Enterpri
 
 * Metronome supports hybrid cloud
 
+* Marathon metrics are now collected by Telegraf (DCOS-47693)
+
 * Expose Public IP (DCOS_OSS-4514)
 
 * Add thisnode.thisdcos.directory dns zone (DCOS_OSS-4666)